model_files/infra-analyzer/requirements.txt
fastapi
uvicorn
groq
python-multipart
python-dotenv
Pillow
requests